WP: restructuring code
