+++ src/gradient_aware_harmonisation/__init__.py
@@ -3,13 +3,18 @@ Gradient-aware harmonisation of timeseries
 """
 
 import importlib.metadata
+import numpy as np
 from typing import Optional, Union
 
 from gradient_aware_harmonisation.timeseries import Timeseries
 from gradient_aware_harmonisation.utils import (
-    ConvergenceMethod,
-    harmonise_splines,
+    add_constant_to_spline,
+    GetHarmonisedSplineLike
 )
+from gradient_aware_harmonisation.convergence import (
+    SplineCosineConvergence
+)
+
 
 __version__ = importlib.metadata.version("gradient_aware_harmonisation")
 
@@ -20,7 +25,8 @@ def harmonise(  # noqa: PLR0913
     harmonisation_time: Union[int, float],
     convergence_timeseries: Timeseries | None = None,
     convergence_time: Optional[Union[int, float]] | None = None,
-    # convergence_method: ConvergenceMethod = ConvergenceMethod.COSINE,
+    get_harmonised_spline: GetHarmonisedSplineLike = SplineCosineConvergence,
+    # get_harmonised_spline: GetHarmonisedSplineLike = get_cosine_decay_harmonised_spline,
     # convergence_function: Callable[[Spline, Spline], Spline] | None = None,
 ) -> Timeseries:
     """
@@ -68,18 +74,50 @@ def harmonise(  # noqa: PLR0913
     if convergence_time is None:
         convergence_time = harmonisee_timeseries.time_axis.max()
 
-    harmonised_spline = harmonise_splines(
-        harmonisee=harmonisee_timeseries.to_spline(),
-        target=target_timeseries.to_spline(),
-        harmonisation_time=harmonisation_time,
-        convergence_spline=convergence_timeseries.to_spline(),
-        convergence_time=convergence_time,
-        #convergence_method=convergence_method,
+    # from timeseries to splines
+    target_spline = target_timeseries.to_spline()
+    harmonisee_spline = harmonisee_timeseries.to_spline()
+    convergence_spline = convergence_timeseries.to_spline()
+
+    # compute derivatives of splines
+    target_dspline = target_spline.derivative()
+    harmonisee_dspline = harmonisee_spline.derivative()
+
+    # match first-order derivatives
+    diff_dspline = np.subtract(
+        target_dspline(harmonisation_time),
+        harmonisee_dspline(harmonisation_time)
     )
 
-    harmonised_timeseries = harmonised_spline(harmonisee_timeseries.time_axis)
+    harmonised_first_derivative = add_constant_to_spline(
+        in_spline=harmonisee_dspline,
+        constant=diff_dspline
+    )
 
-    return harmonised_timeseries
+    # integrate to match zero-order derivative
+    harmonised_spline_first_derivative_only = (
+        harmonised_first_derivative.antiderivative()
+    )
+
+    # match zero-order derivatives
+    diff_spline = np.subtract(
+        target_spline(harmonisation_time),
+        harmonised_spline_first_derivative_only(harmonisation_time)
+    )
+
+    harmonised_spline_no_convergence = add_constant_to_spline(
+        in_spline=harmonised_spline_first_derivative_only,
+        constant=diff_spline
+    )
+
+    harmonised_spline = get_harmonised_spline(
+        harmonisation_time=harmonisation_time,
+        convergence_time=convergence_time,
+        harmonised_spline_no_convergence=harmonised_spline_no_convergence,
+        convergence_spline=convergence_spline,
+    )
+
+    return harmonised_spline(harmonisee_timeseries.time_axis)
 
 
 __all__ = ["harmonise"]
